requirements.txt
langgraph==0.2.44
langchain==0.3.7
langchain-core==0.3.15
langsmith==0.1.137
openai==1.54.4
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2
httpx==0.27.2
requests==2.32.3
# 06-cutting-edge 依赖
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pyyaml==6.0.1
numpy==1.24.3
pillow==10.1.0
asyncio-mqtt==0.16.1
chart-js==4.4.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2